requirements.txt
pytest
requests
faker
pydantic
dotenv
sqlalchemy 
psycopg2
pytz
fastapi
uvicorn
allure-pytest
allure-python-commons
playwright
pytest-playwright